Billing cost by time data refactor (#592)
* REFACOR: move data loading logic into the main page to be used by all charts

* LINT: fix format

* FIX: bug fix for _.uniq issue

* LINT: fix format
